requirements.txt
Django >=1.9.7, <2
PyJWT ~= 1.7.1
django-settingsdict >=1.0.1
oauthlib >=2.0.0
